# Performance tweaks for internal/buffer