### Required for all PRs:

- [ x] Signed [CLA](https://influxdata.com/community/cla/).
- [ ] Associated README.md updated. (if plugin)
- [x ] Has appropriate unit tests.

Benchmark results for internal/buffer
```
Before:
$ go test -run='^$' -bench=.   -benchmem -race
goos: darwin
goarch: amd64
pkg: github.com/influxdata/telegraf/internal/buffer
BenchmarkBufferBatch5Add-8                   	  200000	      5454 ns/op	       2 B/op	       0 allocs/op
BenchmarkBufferBigInfrequentBatchCatchup-8   	 1000000	      1065 ns/op	       1 B/op	       0 allocs/op
BenchmarkBufferOftenBatch-8                  	 1000000	      1892 ns/op	      16 B/op	       0 allocs/op
BenchmarkBufferAlmostBatch-8                 	 1000000	      1543 ns/op	      14 B/op	       0 allocs/op
BenchmarkBufferSlowBatch-8                   	 1000000	      1381 ns/op	       1 B/op	       0 allocs/op
BenchmarkBufferBatchNoDrop-8                 	 1000000	      1792 ns/op	      16 B/op	       0 allocs/op
BenchmarkBufferCatchup-8                     	 1000000	      1230 ns/op	       0 B/op	       0 allocs/op
BenchmarkAddMetrics-8                        	 1000000	      1229 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/influxdata/telegraf/internal/buffer	13.145s

After:
$ go test -run='^$' -bench=.   -benchmem -race
goos: darwin
goarch: amd64
pkg: github.com/influxdata/telegraf/internal/buffer
BenchmarkBufferBatch5Add-8                   	 1000000	      2348 ns/op	       1 B/op	       0 allocs/op
BenchmarkBufferBigInfrequentBatchCatchup-8   	 2000000	       753 ns/op	       1 B/op	       0 allocs/op
BenchmarkBufferOftenBatch-8                  	 2000000	       747 ns/op	      16 B/op	       0 allocs/op
BenchmarkBufferAlmostBatch-8                 	 2000000	       700 ns/op	      14 B/op	       0 allocs/op
BenchmarkBufferSlowBatch-8                   	 2000000	       784 ns/op	       1 B/op	       0 allocs/op
BenchmarkBufferBatchNoDrop-8                 	 2000000	       771 ns/op	      16 B/op	       0 allocs/op
BenchmarkBufferCatchup-8                     	 2000000	       801 ns/op	       0 B/op	       0 allocs/op
BenchmarkAddMetrics-8                        	 2000000	       756 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/influxdata/telegraf/internal/buffer	20.158s


Diff required to make channel based buffer not deadlock:

$ git diff master -- buffer.go
diff --git a/internal/buffer/buffer.go b/internal/buffer/buffer.go
index cdc81fed..174d6632 100644
--- a/internal/buffer/buffer.go
+++ b/internal/buffer/buffer.go
@@ -46,9 +46,14 @@ func (b *Buffer) Add(metrics ...telegraf.Metric) {
                case b.buf <- metrics[i]:
                default:
                        b.mu.Lock()
-                       MetricsDropped.Incr(1)
-                       <-b.buf
-                       b.buf <- metrics[i]
+                       select {
+                       case b.buf <- metrics[i]:
+                       case <-b.buf:
+                               MetricsDropped.Incr(1)
+                               b.buf <- metrics[i]
+                       }
                        b.mu.Unlock()
                }
        }

```