@danielnelson 2018-08-15
@estk Sorry about the delay looking at this PR, can you remove the changes to logger and serializer/influx so that it only has the buffer changes?