@estk 2018-08-15
@danielnelson ill have a chance in the next week.